Implemented save_harmonics_sounds
I implemented save_harmonics_sounds function in io_utils module. The function makes and saves WAV file from the mathematical representation of the signal for each selected instrument. Amplitude of each harmonic is taken into account so that the loudness of each harmonic corresponds to the relative power of the harmonic.